split maze module
