+++ javascript/experiments/maze/src/components/Maze/class/Cell.js
