+++ javascript/experiments/maze/src/components/Maze/class/Edge.js
